Move ForgetTST instances to generic
