+++ lang/syntax/src/trees/generic/exp.rs
@@ -10,7 +10,7 @@ use crate::ctx::values::TypeCtx;
 use crate::ctx::{BindContext, LevelCtx};
 
 use super::traits::Occurs;
-use super::HasTypeInfo;
+use super::{ForgetTST, HasTypeInfo};
 
 #[derive(Debug, Clone, Derivative)]
 #[derivative(Eq, PartialEq, Hash)]
@@ -134,6 +134,22 @@ impl HasTypeInfo for Exp {
     }
 }
 
+impl ForgetTST for Exp {
+    fn forget_tst(&self) -> Self {
+        match self {
+            Exp::Variable(e) => Exp::Variable(e.forget_tst()),
+            Exp::TypCtor(e) => Exp::TypCtor(e.forget_tst()),
+            Exp::Call(e) => Exp::Call(e.forget_tst()),
+            Exp::DotCall(e) => Exp::DotCall(e.forget_tst()),
+            Exp::Anno(e) => Exp::Anno(e.forget_tst()),
+            Exp::TypeUniv(e) => Exp::TypeUniv(e.forget_tst()),
+            Exp::LocalMatch(e) => Exp::LocalMatch(e.forget_tst()),
+            Exp::LocalComatch(e) => Exp::LocalComatch(e.forget_tst()),
+            Exp::Hole(e) => Exp::Hole(e.forget_tst()),
+        }
+    }
+}
+
 // Variable
 //
 //
@@ -188,6 +204,13 @@ impl Occurs for Variable {
     fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
         let Variable { idx, .. } = self;
         ctx.idx_to_lvl(*idx) == lvl
+    }
+}
+
+impl ForgetTST for Variable {
+    fn forget_tst(&self) -> Self {
+        let Variable { span, idx, name, .. } = self;
+        Variable { span: *span, idx: *idx, name: name.clone(), inferred_type: None }
     }
 }
 
@@ -247,6 +270,13 @@ impl Occurs for TypCtor {
     }
 }
 
+impl ForgetTST for TypCtor {
+    fn forget_tst(&self) -> Self {
+        let TypCtor { span, name, args } = self;
+        TypCtor { span: *span, name: name.clone(), args: args.forget_tst() }
+    }
+}
+
 // Call
 //
 //
@@ -299,6 +329,13 @@ impl Occurs for Call {
     fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
         let Call { args, .. } = self;
         args.args.iter().any(|arg| arg.occurs(ctx, lvl))
+    }
+}
+
+impl ForgetTST for Call {
+    fn forget_tst(&self) -> Self {
+        let Call { span, name, args, .. } = self;
+        Call { span: *span, name: name.clone(), args: args.forget_tst(), inferred_type: None }
     }
 }
 
@@ -361,6 +398,19 @@ impl Occurs for DotCall {
     }
 }
 
+impl ForgetTST for DotCall {
+    fn forget_tst(&self) -> Self {
+        let DotCall { span, exp, name, args, .. } = self;
+        DotCall {
+            span: *span,
+            exp: exp.forget_tst(),
+            name: name.clone(),
+            args: args.forget_tst(),
+            inferred_type: None,
+        }
+    }
+}
+
 // Anno
 //
 //
@@ -414,6 +464,12 @@ impl Occurs for Anno {
     }
 }
 
+impl ForgetTST for Anno {
+    fn forget_tst(&self) -> Self {
+        let Anno { span, exp, typ, .. } = self;
+        Anno { span: *span, exp: exp.forget_tst(), typ: typ.forget_tst(), normalized_type: None }
+    }
+}
 // TypeUniv
 //
 //
@@ -452,6 +508,13 @@ impl Shift for TypeUniv {
 impl Occurs for TypeUniv {
     fn occurs(&self, _ctx: &mut LevelCtx, _lvl: Lvl) -> bool {
         false
+    }
+}
+
+impl ForgetTST for TypeUniv {
+    fn forget_tst(&self) -> Self {
+        let TypeUniv { span } = self;
+        TypeUniv { span: *span }
     }
 }
 
@@ -511,6 +574,22 @@ impl Occurs for LocalMatch {
     }
 }
 
+impl ForgetTST for LocalMatch {
+    fn forget_tst(&self) -> Self {
+        let LocalMatch { span, ctx: _, name, on_exp, motive, ret_typ, body, .. } = self;
+        LocalMatch {
+            span: *span,
+            ctx: None,
+            name: name.clone(),
+            on_exp: on_exp.forget_tst(),
+            motive: motive.as_ref().map(|x| x.forget_tst()),
+            ret_typ: ret_typ.as_ref().map(|x| x.forget_tst()),
+            body: body.forget_tst(),
+            inferred_type: None,
+        }
+    }
+}
+
 // LocalComatch
 //
 //
@@ -562,6 +641,21 @@ impl Occurs for LocalComatch {
     }
 }
 
+impl ForgetTST for LocalComatch {
+    fn forget_tst(&self) -> Self {
+        let LocalComatch { span, ctx: _, name, is_lambda_sugar, body, .. } = self;
+
+        LocalComatch {
+            span: *span,
+            ctx: None,
+            name: name.clone(),
+            is_lambda_sugar: *is_lambda_sugar,
+            body: body.forget_tst(),
+            inferred_type: None,
+        }
+    }
+}
+
 // Hole
 //
 //
@@ -605,6 +699,13 @@ impl Occurs for Hole {
     }
 }
 
+impl ForgetTST for Hole {
+    fn forget_tst(&self) -> Self {
+        let Hole { span, .. } = self;
+        Hole { span: *span, inferred_type: None, inferred_ctx: None }
+    }
+}
+
 // Match
 //
 //
@@ -629,6 +730,14 @@ impl Occurs for Match {
     fn occurs(&self, ctx: &mut LevelCtx, lvl: Lvl) -> bool {
         let Match { cases, .. } = self;
         cases.occurs(ctx, lvl)
+    }
+}
+
+impl ForgetTST for Match {
+    fn forget_tst(&self) -> Self {
+        let Match { span, cases, omit_absurd } = self;
+
+        Match { span: *span, cases: cases.forget_tst(), omit_absurd: *omit_absurd }
     }
 }
 
@@ -666,6 +775,19 @@ impl Occurs for Case {
     }
 }
 
+impl ForgetTST for Case {
+    fn forget_tst(&self) -> Self {
+        let Case { span, name, params, body } = self;
+
+        Case {
+            span: *span,
+            name: name.clone(),
+            params: params.forget_tst(),
+            body: body.as_ref().map(|x| x.forget_tst()),
+        }
+    }
+}
+
 // Telescope Inst
 //
 //
@@ -687,6 +809,14 @@ impl TelescopeInst {
     }
 }
 
+impl ForgetTST for TelescopeInst {
+    fn forget_tst(&self) -> Self {
+        let TelescopeInst { params } = self;
+
+        TelescopeInst { params: params.forget_tst() }
+    }
+}
+
 // ParamInst
 //
 //
@@ -703,6 +833,25 @@ pub struct ParamInst {
     pub name: Ident,
     #[derivative(PartialEq = "ignore", Hash = "ignore")]
     pub typ: Option<Rc<Exp>>,
+}
+
+impl Named for ParamInst {
+    fn name(&self) -> &Ident {
+        &self.name
+    }
+}
+
+impl ForgetTST for ParamInst {
+    fn forget_tst(&self) -> Self {
+        let ParamInst { span, name, typ, .. } = self;
+
+        ParamInst {
+            span: *span,
+            info: None,
+            name: name.clone(),
+            typ: typ.as_ref().map(|x| x.forget_tst()),
+        }
+    }
 }
 
 // Args
@@ -737,6 +886,12 @@ impl Shift for Args {
     }
 }
 
+impl ForgetTST for Args {
+    fn forget_tst(&self) -> Self {
+        Args { args: self.args.forget_tst() }
+    }
+}
+
 // Motive
 //
 //
@@ -759,5 +914,13 @@ impl Shift for Motive {
             param: param.clone(),
             ret_typ: ret_typ.shift_in_range(range.shift(1), by),
         }
+    }
+}
+
+impl ForgetTST for Motive {
+    fn forget_tst(&self) -> Self {
+        let Motive { span, param, ret_typ } = self;
+
+        Motive { span: *span, param: param.forget_tst(), ret_typ: ret_typ.forget_tst() }
     }
 }
