+++ lang/syntax/src/trees/tst/mod.rs
@@ -1,4 +1,3 @@
 mod def;
-pub mod forget;
 
 pub use def::*;
