+++ test/test-runner/src/phases.rs
@@ -4,8 +4,7 @@ use std::fmt;
 use parser::cst;
 use printer::PrintToString;
 use renaming::Rename;
-use syntax::tst::forget::ForgetTST;
-use syntax::{tst, ust};
+use syntax::{generic::ForgetTST, tst, ust};
 
 use super::infallible::NoError;
